requirements.txt
django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
markdown==3.3.6
pillow==9.1.0
python-decouple==3.6